Step3( 2등 당첨 추가) pr 요청 (#1048)
* refactor: step 1 피드백 반영
  * utils 클래스 private 생성자 이용. (utils packages)
  * empty,null check에서 불필요한 코드 제거 (validationChecker class)
  * 상수 사용 (Calculator, ValidationChecker class)
  * Pattern 인스턴스 미리 생성처리 (Splitter class)
  * test 메서드에 DisplayName annotation 사용 (CalculatorTest class)
  * 합 계산 로직 stream 사용 (Calculator class)

* docs: step2 구현할 기능 목록 정리

* feat: lotto 랜덤생성 기능 구현.

* test: step1 test코드 수정

* refactor: 로또 생성 기능 수정

* feat: 로또 정렬기능 추가

* feat: 입력된 금액 만큼 로또 생성하는 기능 추가

* refactor: lotto 생성 및 test 코드 리펙토링

* fix: 로또 구매 로직 수정 및 test 추가

* feat: 등수 enum 클래스 생성

* feat: 등수 계산 로직 추가

* feat: lottos 객체 getter 메서드 추가

* feat: 입력받은 winningLotto 처리하는 로직 추가

* refactor: 자료형 변경 (int -> long)

* feat: 필요한 getter 함수 추가

* feat: 결과 계산 로직 추가

* feat: main, view 처리 로직 추가

* refactor: 예외처리를 위한 형변환

* feat: 유효성검증 로직 추가

* feat: 비지니스 로직에 유효성검사 & test 추가

* refactor: public -> private 변경

* refactor: 피드백 반영
* 로또 일치하는 갯수 계산하는 로직 추가 (lotto)
* Set 인스턴스 변수 추가 (lotto)
* Number 객체 이용해서 Integer wrapping함 (lotto)
* 입력받은 winninglotto string을 처리(split)로직 이동 (lottoAdmin -> lotto)
* Numbers 클래스 이름 변경
* lottos 클래스 삭제

* refactor: 피드백 반영
* 결과 계산로직 변경, map 자료구조 사용X (result)
* lotto 생성시 Number 외부주입 제거 (shop)
* 유효성 검사 클래스 수정

* refactor: main, outputView 수정

* refactor: 오류 수정

* refactor: 피드백 반영, 오류 수정
* return type (int -> double)

* refactor: 피드백 반영, 오류 수정
* LottoAdmin 제거
* 결과 계산 로직 이동 (LottoAdmin -> Lotto)
* 관련 test 수정

* test: 피드백 반영
* 결과 계산 로직 test 수정

* refactor: 피드백 반영
* money 입력값 변경 (String -> long)

* refactor: 피드백 반영
* List 인스턴스 변수 제거 (정합성에 대한 load 감소)
* shuffle & 6자리 추출 로직 삽입 (기존: lotto -> 변경: Number)
* 객체 내에서 null 예외처리 추가
* 기존 validationCheck 기능 삭제

* refactor: 피드백 반영
* 구매한 lotto를 보여줄때 list 로 형변환 하여 정렬한뒤 보여주도록 변경

* fix: step2 피드백 반영: 자료형변경 (double -> long)

* docs: step3 기능목록 정리

* fix: 소숫점 표시를 위한 casting

* feat: 2등 enum에 추가, 계산로직 추가, 관련 test 추가

* feat: 2등에 대한 입출력 추가

+++ src/main/java/lotto/domain/Result.java
@@ -18,11 +18,11 @@ public class Result {
         return result;
     }
 
-    public double getTotalWinningPrice() {
-        return result.stream().mapToDouble(rank -> rank.getWinningMoney()).sum();
+    public long getTotalWinningPrice() {
+        return result.stream().mapToLong(rank -> rank.getWinningMoney()).sum();
     }
 
     public String getWinningRatio(long purchasedPrice) {
-        return String.format(DIGIT, getTotalWinningPrice() / purchasedPrice);
+        return String.format(DIGIT, getTotalWinningPrice() / (double) purchasedPrice);
     }
 }
